refactor: move update_nested_dict to common module
