+++ src/ai/backend/accelerator/cuda_open/plugin.py
@@ -28,8 +28,8 @@ from ai.backend.agent.resources import (
     DeviceSlotInfo,
     DiscretePropertyAllocMap,
 )
-from ai.backend.agent.utils import update_nested_dict
 from ai.backend.common.logging import BraceStyleAdapter
+from ai.backend.common.utils import update_nested_dict
 
 try:
     from ai.backend.agent.resources import get_resource_spec_from_container  # type: ignore
