+++ src/ai/backend/agent/kubernetes/utils.py
@@ -10,8 +10,7 @@ from aiodocker.docker import Docker
 from aiodocker.exceptions import DockerError
 
 from ai.backend.common.logging import BraceStyleAdapter
-
-from ..utils import update_nested_dict
+from ai.backend.common.utils import update_nested_dict
 
 log = BraceStyleAdapter(logging.getLogger(__spec__.name))  # type: ignore[name-defined]
